+++ src/lib/components/nav/SettingsButton.svelte
@@ -2,7 +2,6 @@
   import Button from '$components/ui/Button.svelte';
   import Settings from '../modals/Settings.svelte';
   import * as Dialog from '$lib/components/ui/dialog';
-  import { portal } from '$stores/portal';
 
   let settingsDialogOpen: boolean = false;
 
@@ -11,7 +10,7 @@
   };
 </script>
 
-<Dialog.Root portal={$portal} bind:open={settingsDialogOpen}>
+<Dialog.Root bind:open={settingsDialogOpen}>
   <Dialog.Trigger asChild let:builder>
     <Button builders={[builder]} variant="ghost">Settings</Button>
   </Dialog.Trigger>