+++ src/lib/components/ui/dialog/index.ts
@@ -1,5 +1,6 @@
 import { Dialog as DialogPrimitive } from "bits-ui";
 
+import Root from "./dialog-root.svelte";
 import Title from "./dialog-title.svelte";
 import Portal from "./dialog-portal.svelte";
 import Footer from "./dialog-footer.svelte";
@@ -8,7 +9,7 @@ import Overlay from "./dialog-overlay.svelte";
 import Content from "./dialog-content.svelte";
 import Description from "./dialog-description.svelte";
 
-const Root = DialogPrimitive.Root;
+// const Root = DialogPrimitive.Root;
 const Trigger = DialogPrimitive.Trigger;
 const Close = DialogPrimitive.Close;