+++ src/lib/components/ui/dropdown-menu/dropdown-menu-root.svelte
@@ -1,16 +1,10 @@
 <script lang="ts">
-  import { DropdownMenu as DropdownMenuPrimitive } from 'bits-ui';
+  import { DropdownMenu as DropdownMenuPrimitive, type DropdownMenuProps} from 'bits-ui';
   import { portal } from '$stores/portal';
-  import { onMount } from 'svelte';
 
-  type $$Props = DropdownMenuPrimitive.Props;
-
-  onMount(() => {
-    console.log($portal);
-  });
+  type $$Props = DropdownMenuProps;
 </script>
 
-<DropdownMenuPrimitive.Root portal={null} {...$$restProps}>
-<!-- <DropdownMenuPrimitive.Root portal={$portal} {...$$restProps} > -->
+<DropdownMenuPrimitive.Root portal={$portal} {...$$restProps}>
   <slot />
 </DropdownMenuPrimitive.Root>