+++ src/lib/components/ui/popover/index.ts
@@ -1,6 +1,7 @@
 import { Popover as PopoverPrimitive } from "bits-ui";
 import Content from "./popover-content.svelte";
-const Root = PopoverPrimitive.Root;
+import Root from "./popover-root.svelte";
+
 const Trigger = PopoverPrimitive.Trigger;
 const Close = PopoverPrimitive.Close;