+++ src/lib/components/ui/tooltip/tooltip-root.svelte
@@ -1,0 +1,10 @@
+<script lang="ts">
+  import { Tooltip as TooltipPrimitive } from 'bits-ui';
+  import { portal } from '$stores/portal';
+
+  type $$Props = TooltipPrimitive.Props;
+</script>
+
+<TooltipPrimitive.Root portal={$portal}>
+  <slot />
+</TooltipPrimitive.Root>